catch up with the main site.

+++ tests/test_binary_flat.cpp
@@ -1,8 +1,7 @@
 /**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
- * This source code is licensed under the BSD+Patents license found in the
+ * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
@@ -35,7 +34,7 @@ TEST(BinaryFlat, accuracy) {
     index.add(nb, database.data());
   }
 
-  int nq = 200;
+  size_t nq = 200;
 
   { // searching the database
